fix(mpv): keep host-integrated Linux libraries external
Amp-Thread-ID: https://ampcode.com/threads/T-019f99dc-468a-74fb-9fb0-d69623be5ddc
